Insurance Claim Analysis
The given problem has financial impact on the organization. The objective of the 
organization through this analysis is to identify the fraudulent claims. 
This project will support the Claims team to effectively manage the claims through 
predicting whether the submitted claim is to be accepted or rejected. Any claim 
rejected by the model can be scrutinized by the team in detail and hence, take a 
final call whether to accept the claim or reject the same.